requirement.txt
attrs==25.3.0
certifi==2025.8.3
dill==0.4.0
h11==0.16.0
idna==3.10
isort==6.0.1
mccabe==0.7.0
outcome==1.3.0.post0
platformdirs==4.3.8
PySocks==1.7.1
python-dotenv==1.1.1
selenium==4.34.2
sniffio==1.3.1
sortedcontainers==2.4.0
tomlkit==0.13.3
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.14.1
urllib3==2.5.0
websocket-client==1.8.0
wsproto==1.2.0
